Initial source files+gitignore
I hope I set the gitignore up correctly, idea is you can take the normal
GENIUS project you already have and copy that to this folder. All files
except this set are ignored, so if you want to create new seperate
source files add them as exceptions to the gitignore.